add mutex to toast struct
