+++ ui/notification/system.go
@@ -1,6 +1,9 @@
 package notification
 
 import (
+	"fmt"
+	"os"
+	"path"
 	"path/filepath"
 
 	"github.com/gen2brain/beeep"
@@ -32,4 +35,18 @@ func (s *SystemNotification) Notify(message string) {
 	if err != nil {
 		log.Info("could not initiate desktop notification, reason:", err.Error())
 	}
+}
+
+func getAbsolutePath() (string, error) {
+	ex, err := os.Executable()
+	if err != nil {
+		return "", fmt.Errorf("error getting executable path: %s", err.Error())
+	}
+
+	exSym, err := filepath.EvalSymlinks(ex)
+	if err != nil {
+		return "", fmt.Errorf("error getting filepath after evaluating sym links")
+	}
+
+	return path.Dir(exSym), nil
 }
