--- ui/notification/utils.go
@@ -1,22 +1,0 @@
-package notification
-
-import (
-	"fmt"
-	"os"
-	"path"
-	"path/filepath"
-)
-
-func getAbsolutePath() (string, error) {
-	ex, err := os.Executable()
-	if err != nil {
-		return "", fmt.Errorf("error getting executable path: %s", err.Error())
-	}
-
-	exSym, err := filepath.EvalSymlinks(ex)
-	if err != nil {
-		return "", fmt.Errorf("error getting filepath after evaluating sym links")
-	}
-
-	return path.Dir(exSym), nil
-}
